Enable controller board and Win9x playback
The controller board polling task on core 0 was disabled for
development. It has been enabled.

Windows 9x requires the ATAPI CDB Seek 10 to stop playback and return
an OK status. This has been partially implemented, it doesn't actually
perform a seek but it does stop playback so Windows is happy.
Seek isn't used often because read, write, an audio playback commands
already seek to the start of their operation.

Also adjusted the frequency divider of the SD card initialization.
Due to the increase in the sys clk frequency the divider down to 1MHz
for the SDIO clock had to be increased.